refactor: 쿠폰 도메인 엔티티 및 enums 리팩토링 - models 디렉토리로 이동
